Adds the ability to set the threshold in config
Closes #5

This is a non-breaking change, just expands what the config
file can do. The readme includes the details of the change
and an example of how to adjust your config.

The default threshold for including someone's email in the
result, and adding them to a review with publish, remains
20% of the file belonging to them. However, this patch lets
you set that to any percentage you would like through the
`threshold` option in the config file.

Impact: Low
Bump: Minor