fix excluded directory trailing slash bug
Prior to this patch there was a bug in the excluded directory check due
to `path.dir` stripping the trailing slash off of the examined file. It
could be fixed by not including the trailing slash in the configuration,
but that would allow an excluded directory of `foo/bar` to also match
`foo/baroops`.

This patch appends a trailing slash to `path.dir` and matches against
that. It also breaks the method out into it's own file so we can test
it.